+++ packages/Toast/Growl.tsx
@@ -2,10 +2,9 @@ import React, { Children, cloneElement } from 'react'
 import { ClearButton } from '@welcome-ui/clear-button'
 import { CreateWuiProps, forwardRef } from '@welcome-ui/system'
 
-import { Title } from './Title'
 import * as S from './styles'
 
-import { Variant } from '.'
+import { Toast, Variant } from './index'
 
 export interface GrowlOptions {
   variant?: Variant
@@ -23,7 +22,7 @@ export const Growl = forwardRef<'div', GrowlProps>(
   ({ children, closeButtonDataTestId, hasCloseButton = true, onClose, variant = 'info' }, ref) => {
     const content = Children.map(children, child => {
       // Add variant to AlertTitle to show the correct icon/color
-      if ((child as React.ReactElement).type === Title) {
+      if ((child as React.ReactElement).type === Toast.Title) {
         return cloneElement(child as React.ReactElement, { variant })
       }
       return child
