+++ packages/Toast/Snackbar.tsx
@@ -2,10 +2,9 @@ import React from 'react'
 import { ClearButton } from '@welcome-ui/clear-button'
 import { CreateWuiProps, forwardRef } from '@welcome-ui/system'
 
-import { Title, TitleOptions } from './Title'
 import * as S from './styles'
 
-import { Variant } from '.'
+import { TitleOptions, Toast, Variant } from './index'
 
 export interface SnackbarOptions {
   variant?: Variant
@@ -21,12 +20,12 @@ export type SnackbarProps = CreateWuiProps<'div', SnackbarOptions & TitleOptions
 export const Snackbar = forwardRef<'div', SnackbarProps>(
   ({ children, hasCloseButton = true, icon, onClose, variant = 'info', ...rest }, ref) => (
     <S.Snackbar hasCloseButton={hasCloseButton} ref={ref} variant={variant} {...rest}>
-      <Title icon={icon} variant={variant}>
+      <Toast.Title icon={icon} variant={variant}>
         <>
           {children}
           {hasCloseButton && <ClearButton onClick={onClose} />}
         </>
-      </Title>
+      </Toast.Title>
     </S.Snackbar>
   )
 )
